+++ src/main/server/player/webview.ts
@@ -10,6 +10,13 @@ const SessionKeys = {
 };
 
 export function useWebview(player: alt.Player) {
+    /**
+     * Emit an event directly to the webview
+     *
+     * @param {string} eventName
+     * @param {...any[]} args
+     * @return
+     */
     function emit(eventName: string, ...args: any[]) {
         if (!player || !player.valid) {
             return;
@@ -18,14 +25,32 @@ export function useWebview(player: alt.Player) {
         player.emit(Events.view.onServer, eventName, ...args);
     }
 
-    function show(vuePage: PageNames, type: PageType) {
+    /**
+     * Show a specific webview page to the player.
+     *
+     * Assign the `type` of page to change how it behaves.
+     *
+     * Escape to close page, only effects page types.
+     *
+     * @param {PageNames} vuePage
+     * @param {PageType} type
+     * @param {boolean} [escapeToClosePage=false]
+     * @return
+     */
+    function show(vuePage: PageNames, type: PageType, escapeToClosePage = false) {
         if (!player || !player.valid) {
             return;
         }
 
-        player.emit(Events.view.show, vuePage, type);
+        player.emit(Events.view.show, vuePage, type, escapeToClosePage);
     }
 
+    /**
+     * Hide a page that was shown to the player
+     *
+     * @param {PageNames} vuePage
+     * @return
+     */
     function hide(vuePage: PageNames) {
         if (!player || !player.valid) {
             return;
@@ -34,6 +59,11 @@ export function useWebview(player: alt.Player) {
         player.emit(Events.view.hide, vuePage, 'page');
     }
 
+    /**
+     * Focus the webview, and show the cursor
+     *
+     * @return
+     */
     function focus() {
         if (!player || !player.valid) {
             return;
@@ -42,6 +72,11 @@ export function useWebview(player: alt.Player) {
         player.emit(Events.view.focus);
     }
 
+    /**
+     * Unfocus the webview, and hide the cursor
+     *
+     * @return
+     */
     function unfocus() {
         if (!player || !player.valid) {
             return;
@@ -50,6 +85,13 @@ export function useWebview(player: alt.Player) {
         player.emit(Events.view.unfocus);
     }
 
+    /**
+     * Verify if a page is currently ready
+     *
+     * @param {PageNames} pageName
+     * @param {PageType} type
+     * @return
+     */
     async function isReady(pageName: PageNames, type: PageType) {
         if (type === 'page') {
             try {